cli: move code between aks.ts and util.ts
